Updates Kotha (2019 - now 2020) model (excl. slope/geology)
