+++ openquake/hazardlib/gsim/sera_amplification_models.py
@@ -28,7 +28,6 @@ from scipy.constants import g
 # from scipy.interpolate import interp1d
 from openquake.hazardlib.gsim.base import (GMPE, CoeffsTable, registry)
 from openquake.hazardlib.imt import PGA, SA, from_string
-from openquake.hazardlib.gsim.kotha_2019 import KothaEtAl2019SERA
 from openquake.hazardlib import const
 
 
@@ -37,27 +36,27 @@ imls = [0., 0.25, 0.5, 0.75, 1., 1.25]
 
 # Short period amplification factors defined by Pitilakis et al., (2018)
 FS = {
-"A":  [1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
-"B1": [1.3, 1.3, 1.2, 1.2, 1.2, 1.2],
-"B2": [1.4, 1.3, 1.3, 1.2, 1.1, 1.1],
-"C1": [1.7, 1.6, 1.4, 1.3, 1.3, 1.2],
-"C2": [1.6, 1.5, 1.3, 1.2, 1.1, 1.0],
-"C3": [1.8, 1.6, 1.4, 1.2, 1.1, 1.0],
-"D":  [2.2, 1.9, 1.6, 1.4, 1.2, 1.0],
-"E":  [1.7, 1.6, 1.6, 1.5, 1.5, 1.5]
+    "A":  [1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
+    "B1": [1.3, 1.3, 1.2, 1.2, 1.2, 1.2],
+    "B2": [1.4, 1.3, 1.3, 1.2, 1.1, 1.1],
+    "C1": [1.7, 1.6, 1.4, 1.3, 1.3, 1.2],
+    "C2": [1.6, 1.5, 1.3, 1.2, 1.1, 1.0],
+    "C3": [1.8, 1.6, 1.4, 1.2, 1.1, 1.0],
+    "D":  [2.2, 1.9, 1.6, 1.4, 1.2, 1.0],
+    "E":  [1.7, 1.6, 1.6, 1.5, 1.5, 1.5]
 }
 
 
 # Long period amplification factors defined by Pitilakis et al., (2018)
 F1 = {
-"A":  [1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
-"B1": [1.4, 1.4, 1.4, 1.4, 1.3, 1.3],
-"B2": [1.6, 1.5, 1.5, 1.5, 1.4, 1.3],
-"C1": [1.7, 1.6, 1.5, 1.5, 1.4, 1.3],
-"C2": [2.1, 2.0, 1.9, 1.8, 1.8, 1.7],
-"C3": [3.2, 3.0, 2.7, 2.5, 2.4, 2.3],
-"D":  [4.1, 3.8, 3.3, 3.0, 2.8, 2.7],
-"E":  [1.3, 1.3, 1.2, 1.2, 1.2, 1.2]
+    "A":  [1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
+    "B1": [1.4, 1.4, 1.4, 1.4, 1.3, 1.3],
+    "B2": [1.6, 1.5, 1.5, 1.5, 1.4, 1.3],
+    "C1": [1.7, 1.6, 1.5, 1.5, 1.4, 1.3],
+    "C2": [2.1, 2.0, 1.9, 1.8, 1.8, 1.7],
+    "C3": [3.2, 3.0, 2.7, 2.5, 2.4, 2.3],
+    "D":  [4.1, 3.8, 3.3, 3.0, 2.8, 2.7],
+    "E":  [1.3, 1.3, 1.2, 1.2, 1.2, 1.2]
 }
 
 
@@ -519,7 +518,7 @@ class SandikkayaDinsever2018(GMPE):
         stddev_check = (const.StdDev.INTER_EVENT in
                         self.DEFINED_FOR_STANDARD_DEVIATION_TYPES) and\
                         (const.StdDev.INTRA_EVENT in
-                        self.DEFINED_FOR_STANDARD_DEVIATION_TYPES)
+                         self.DEFINED_FOR_STANDARD_DEVIATION_TYPES)
         if not stddev_check:
             raise ValueError("Input GMPE %s not defined for inter- and intra-"
                              "event standard deviation" % str(self.gmpe))
@@ -544,7 +543,6 @@ class SandikkayaDinsever2018(GMPE):
                                  % " ".join(REGION_SET))
         else:
             self.region = region
-
 
     def get_mean_and_stddevs(self, sctx, rctx, dctx, imt, stddev_types):
         """
